New port: devel/py-subprocess32
Backport of the subprocess module from Python 3.2/3.3 for use on
2.x.

This is a backport of the subprocess standard library module from
Python 3.2 & 3.3 for use on Python 2.4, 2.5, 2.6 and 2.7. It includes
bugfixes and new features. On POSIX systems it is guaranteed to be
reliable when used in threaded applications. Bonus: It includes
timeout support from Python 3.3.

WWW: https://pypi.python.org/pypi/subprocess32/